Handle track selection
